fix: unused import and functions
